Write output to buffer instead of file
Instead of writing the generated docx contents to a temporary file,
we now just create a string buffer with the file content.
This makes it easier to use the gem on a system where the application might not
have write access to the filesystem. It also ensures that the gem doesn't leave
behind a lot of temporary files, since there was no system for cleaning them up
after use.

This is also the recommended way of using Rubyzip for modifying docx files:
https://github.com/rubyzip/rubyzip#modify-docx-file-with-rubyzip