add image types to content types
Relevant image content types must be included in [Content_Types].xml

Currently this just replaces the entire file with a basic one including
all common content types.

It would be better to only include those that actually are present, and
even better if we just appended to content_types file those that are
required and don’t already exist.